[Optimize] make `encoder.h` deps standalone
As `encoder.h` may be exposed to developer to invoke method inside, we need to make sure that `encoder.h` does not deps
any class that should not be exposed.

issue: f-23904875
AutoSubmit: true
